+++ outsystems_integrations/architecture_dashboard/vars.py
@@ -1,0 +1,4 @@
+# Architecture Dashboard API Integration Variables
+# Sync specific
+SLEEP_PERIOD_IN_SECS = 1800
+MAX_RETRIES = 3